Moved getEventDetails & getEventType to another file
