+++ frontend/src/modules/app/EventAlerts.tsx
@@ -4,8 +4,6 @@ import { shallowEqual, useDispatch, useSelector } from 'react-redux';
 import { Button, Grid, makeStyles, Theme, Typography } from '@material-ui/core';
 import VolumeOffIcon from '@material-ui/icons/VolumeOff';
 import VolumeUpIcon from '@material-ui/icons/VolumeUp';
-import { LogEventCode, LogEventType } from '../../store/controller/proto/mcu_pb';
-import { BMIN, BPM, LMIN, PERCENT } from '../info/units';
 import {
   getActiveLogEventIds,
   getAlarmMuteStatus,
@@ -15,7 +13,8 @@ import ModalPopup from '../controllers/ModalPopup';
 import LogsPage from '../logs/LogsPage';
 import { BellIcon } from '../icons';
 import { updateCommittedState } from '../../store/controller/actions';
-import { ALARM_MUTE, BACKEND_CONNECTION_LOST_CODE } from '../../store/controller/types';
+import { ALARM_MUTE } from '../../store/controller/types';
+import { getEventType } from '../logs/EventType';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
@@ -97,110 +96,6 @@ const useStyles = makeStyles((theme: Theme) => ({
     marginRight: '10px',
   },
 }));
-
-export interface EventType {
-  type: LogEventType;
-  label: string;
-  unit: string;
-  head?: string;
-  stateKey?: string;
-}
-export const getEventType = (code: LogEventCode): EventType => {
-  switch (code) {
-    case LogEventCode.fio2_too_low:
-      return {
-        type: LogEventType.patient,
-        label: 'fiO2 is too low',
-        head: 'FiO2',
-        stateKey: 'fio2',
-        unit: PERCENT,
-      };
-    case LogEventCode.fio2_too_high:
-      return {
-        type: LogEventType.patient,
-        label: 'fiO2 is too high',
-        head: 'FiO2',
-        stateKey: 'fio2',
-        unit: PERCENT,
-      };
-    case LogEventCode.rr_too_low:
-      return {
-        type: LogEventType.patient,
-        label: 'Respiratory Rate is too low',
-        stateKey: 'rr',
-        unit: BMIN,
-      };
-    case LogEventCode.rr_too_high:
-      return {
-        type: LogEventType.patient,
-        label: 'Respiratory Rate is too high',
-        stateKey: 'rr',
-        unit: BMIN,
-      };
-    case LogEventCode.hr_too_low:
-      return {
-        type: LogEventType.patient,
-        label: 'Heart Rate is too low',
-        stateKey: 'hr',
-        unit: BPM,
-      };
-    case LogEventCode.hr_too_high:
-      return {
-        type: LogEventType.patient,
-        label: 'Heart Rate is too high',
-        stateKey: 'hr',
-        unit: BPM,
-      };
-    case LogEventCode.spo2_too_low:
-      return {
-        type: LogEventType.patient,
-        label: 'spO2 is too low',
-        stateKey: 'spo2',
-        unit: PERCENT,
-      };
-    case LogEventCode.spo2_too_high:
-      return {
-        type: LogEventType.patient,
-        label: 'spO2 is too high',
-        stateKey: 'spo2',
-        unit: PERCENT,
-      };
-    case LogEventCode.fio2_setting_changed:
-      return {
-        type: LogEventType.control,
-        label: 'Fio2 Settings changed',
-        stateKey: 'fio2',
-        unit: PERCENT,
-      };
-    case LogEventCode.flow_setting_changed:
-      return {
-        type: LogEventType.control,
-        label: 'Flow Settings changed',
-        stateKey: 'flow',
-        unit: LMIN,
-      };
-    case LogEventCode.battery_low:
-      return {
-        type: LogEventType.system,
-        label: 'Battery power is low',
-        unit: PERCENT,
-      };
-    case LogEventCode.screen_locked:
-      return {
-        type: LogEventType.system,
-        label: 'Screen is locked',
-        unit: '',
-      };
-    case BACKEND_CONNECTION_LOST_CODE:
-      return {
-        type: LogEventType.system,
-        label: 'Software connectivity lost',
-        unit: '',
-      };
-    default:
-      return { type: LogEventType.system, label: '', unit: '' };
-  }
-};
 
 interface Props {
   label: string;
